refactor(api): reorganize API handlers into dedicated modules
- Removed the old handlers file and created separate modules for message, image, speak, and whisper handlers, enhancing clarity and maintainability.
- Updated the API router to utilize the new handler functions, adhering to the single-responsibility principle.
- Introduced helper functions for common operations, improving code reuse and organization.
