+++ packages/client-direct/src/handlers/index.ts
@@ -1,1 +1,5 @@
 export * as getRequests from "./getRequests";
+export * as image from "./image";
+export * as message from "./message";
+export * as speak from "./speak";
+export * as whisper from "./whisper";
